Fix options(), minimax() and evaluate()
canMove list in options() contained errors and was incomplete.

minimax() calls itself as the other player now and returns the opposite
of the best score it found each recursion. That's because each recursion
the opponent is at turn and a good score for one player is a bad score
for the other.

evaluate() now looks at which player is at turn so both player have to
maximize the score.